Debugging
Fixed known bugs, all functions now work as they should.

Deleted some functions and methods that were meant to prettify the UI on the ethereum wallet, since site is being built, it should be handled on the front-end instead. These were stripped to make sufficient room for the additional ESOP functions to come.